added constructor for PG class

+++ .github/Random.java
@@ -17,7 +17,7 @@ public class Random {
         return "";
     }
 
-    public static int getRandomInt(double min, double max){
+    public static int getRandomInt(int min, int max){
         return (int)(Math.floor(Math.random()*(max-min)+min));
     }
 }
